+++ packages/nutui-taro-demo-rn/scripts/taro/adapted.js
@@ -1,1 +1,2 @@
-exports = module.exports = ["cell","cellgroup","overlay","button","switch","image"];
+exports = module.exports = ["cell","cellgroup","overlay","button","switch","countdown","empty","price","tag","space", "image"];
+