+++ src/packages/cellgroup/cellgroup.tsx
@@ -40,8 +40,10 @@ export const CellGroup: FunctionComponent<Partial<CellGroupProps>> = (
       >
         <CellGroupContext.Provider value={{ divider, group: true }}>
           {React.Children.map(children, (child, index) => {
+            // @ts-ignore
             return child?.type?.displayName === 'NutCell'
-              ? React.cloneElement(child, {
+              ? // @ts-ignore
+                React.cloneElement(child, {
                   isLast: index === React.Children.count(children) - 1,
                 })
               : child
@@ -52,5 +54,4 @@ export const CellGroup: FunctionComponent<Partial<CellGroupProps>> = (
   )
 }
 
-CellGroup.defaultProps = defaultProps
 CellGroup.displayName = 'NutCellGroup'