+++ src/packages/indicator/indicator.taro.tsx
@@ -23,6 +23,8 @@ export const Indicator: FunctionComponent<
   const classes = classNames({
     [`${classPrefix}-vertical`]: direction === 'vertical',
   })
+  const classPrefixV =
+    direction === 'vertical' ? `${classPrefix}-vertical` : classPrefix
   const renderElement = () => {
     const childs: ReactNode[] = []
     for (let item = 0; item < total; item++) {
@@ -31,11 +33,14 @@ export const Indicator: FunctionComponent<
           children || (
             <View
               key={item}
-              className={`${classPrefix}-dot ${classPrefix}-active`}
+              className={`${classPrefix}-dot ${classPrefix}-active ${classPrefixV}-dot ${classPrefixV}-active`}
             />
           )
         ) : (
-          <View key={item} className={`${classPrefix}-dot`} />
+          <View
+            key={item}
+            className={`${classPrefix}-dot ${classPrefixV}-dot`}
+          />
         )
       )
     }
@@ -48,5 +53,4 @@ export const Indicator: FunctionComponent<
   )
 }
 
-Indicator.defaultProps = defaultProps
 Indicator.displayName = 'NutIndicator'