+++ src/packages/safearea/safearea.taro.tsx
@@ -18,3 +18,5 @@ export const SafeArea: FC<SafeAreaProps> = (props) => {
     />
   )
 }
+
+SafeArea.displayName = 'NutSafeArea'