+++ src/packages/sidenavbar/sidenavbar.tsx
@@ -34,12 +34,13 @@ export const SideNavBar: FunctionComponent<Partial<SideNavBarProps>> = (
     children,
     className,
     onClose,
+    indent,
     ...rest
   } = {
     ...defaultProps,
     ...props,
   }
-  const indent = props.indent ? Number(props.indent) : 20
+  const innerIndent = indent ? Number(indent) : 20
 
   return (
     <Popup
@@ -56,11 +57,11 @@ export const SideNavBar: FunctionComponent<Partial<SideNavBarProps>> = (
           >
             <div
               className={`${classPrefix}-title ${classPrefix}-border-bt`}
-              style={{ paddingLeft: `${indent}px` }}
+              style={{ paddingLeft: `${innerIndent}px` }}
             >
               {title} <i className="arrow-icon arrow-down" />
             </div>
-            <OffsetContext.Provider value={indent}>
+            <OffsetContext.Provider value={innerIndent}>
               <div className={`${classPrefix}-content`}>{children}</div>
             </OffsetContext.Provider>
           </div>
@@ -70,5 +71,4 @@ export const SideNavBar: FunctionComponent<Partial<SideNavBarProps>> = (
   )
 }
 
-SideNavBar.defaultProps = defaultProps
 SideNavBar.displayName = 'NutSideNavBar'