+++ src/packages/step/step.tsx
@@ -75,5 +75,4 @@ export const Step: FunctionComponent<
   )
 }
 
-Step.defaultProps = defaultProps
 Step.displayName = 'NutStep'