+++ src/packages/sticky/sticky.taro.tsx
@@ -165,5 +165,4 @@ export const Sticky: FunctionComponent<Partial<StickyProps>> = (props) => {
   )
 }
 
-Sticky.defaultProps = defaultProps
 Sticky.displayName = 'NutSticky'